+++ src/lib/index.js
@@ -58,6 +58,12 @@ import LoadingOverlay from "./components/LoadingOverlay.react";
 import Avatar from "./components/Avatar.react";
 import AvatarGroup from "./components/AvatarGroup.react";
 import Kbd from "./components/Kbd.react";
+import Highlight from "./components/Highlight.react";
+import Timeline from "./components/Timeline.react";
+import TimelineItem from "./components/TimelineItem.react";
+import Menu from "./components/Menu.react";
+import MenuItem from "./components/MenuItem.react";
+import MenuLabel from "./components/MenuLabel.react";
 import Popover from "./components/Popover.react";
 
 export {
@@ -121,5 +127,11 @@ export {
     Avatar,
     AvatarGroup,
     Kbd,
+    Highlight,
+    Timeline,
+    TimelineItem,
+    Menu,
+    MenuItem,
+    MenuLabel,
     Popover,
 };