+++ CHANGELOG.md
@@ -9,10 +9,12 @@
   - Avatar, AvatarGroup
   - Highlight
   - Timeline, TimelineInput
+  - Mark
 
 ### Fixed
 
-- DateRangePicker will keep the half selected range on dropdown close. # 47
+- DateRangePicker will keep the half selected range on dropdown close. #47
+- DatePicker and DateRangePicker timezone issue. #33
 
 # 0.5.0